Creating __init__ file, separating into modules
1) Created a __init__ file that defines the App and the DB variables.
2) Separated the models into new file (models.py).
3) Reorganized the controllers.py a little.
