+++ flask_app.py
@@ -1,33 +1,7 @@
-from flask import Flask, request, render_template, render_template_string
-from flask_sqlalchemy import SQLAlchemy
+from __init__ import app, db
+from models import Tasks
+from flask import request, render_template_string, render_template
 from controllers import static_html_string, creating_table_rows_string, create_full_html_string, get_all_tasks
-
-app=Flask('__name__')
-#Changing some basic configurations in the Application.
-app.config.update(
-	SECRET_KEY='topsecret',
-	SQLALCHEMY_DATABASE_URI='sqlite:////Users/Tomer Ben-Levi/Projects/ToDoList/example.db',
-	SQLALCHEMY_TRACK_MODIFICATIONS=False)
-
-db=SQLAlchemy(app)
-
-class Tasks(db.Model):
-	"""
-	This class creates the Tasks table in the DB.
-	There are 3 clomuns here: Task_ID(id), Task_Description(task) and a mark whether the task is done or not (isdone).
-	"""
-	__tablename__ = 'Tasks'
-
-	id=db.Column(db.Integer, primary_key=True)
-	task=db.Column(db.String(255), nullable=False)
-	isdone=db.Column(db.String(1), nullable=True)
-
-	def __init__(self, task=None, isdone='0'):
-		self.task=task
-		self.isdone=isdone
-	
-	def __repr__(self):
-		return 'The id is {}, Task is {}'.format(self.id, self.task)
 
 @app.route('/homepage', methods=['GET'])
 def homepage():
@@ -42,7 +16,7 @@ def create_task(class_name=Tasks):
 	NewTask = class_name(query_task)
 	db.session.add(NewTask)
 	db.session.commit()
-	return render_template('homepage.html')
+	return render_template_string(homepage())
 
 @app.route('/removing', methods=['DELETE'])
 def remove_task(class_name=Tasks):
@@ -64,3 +38,10 @@ def update_task(class_name=Tasks):
 if __name__=='__main__':
 	db.create_all()
 	app.run(debug=True, use_reloader=False)
+
+"""
+WHAT'S NEXT:
+- Fix the bug with the creating page that opens after adding a task and gets stuck.
+- Add the Remove and Update options - text field with task number maybe.
+- Use Git.
+"""
